Move upper_active functions from prepare to select module and remove upper_active_layer function as it is duplicate of get_upper_active_cells.
