+++ imod/select/__init__.py
@@ -4,7 +4,12 @@ Get points, cross sections and layers.
 
 from imod.select.cross_sections import cross_section_line, cross_section_linestring
 from imod.select.grid import active_grid_boundary_xy, grid_boundary_xy
-from imod.select.layers import upper_active_layer
+from imod.select.layers import (
+    get_lower_active_grid_cells,
+    get_lower_active_layer_number,
+    get_upper_active_grid_cells,
+    get_upper_active_layer_number,
+)
 from imod.select.points import (
     points_in_bounds,
     points_indices,
