+++ imod/select/layers.py
@@ -1,41 +1,74 @@
-def upper_active_layer(da, is_ibound=True, include_constant_head=False):
+"""
+This module contains all kinds of utilities to work with layers.
+"""
+
+from imod.typing import GridDataArray
+from imod.typing.grid import preserve_gridtype
+
+
+@preserve_gridtype
+def get_upper_active_layer_number(active: GridDataArray) -> GridDataArray:
     """
-    Function to get the upper active layer from ibound xarray.DataArray
+    Returns planar grid of integers with the layer number of the upper most
+    active cell.
 
     Parameters
     ----------
-    da : xarray.DataArray
-        A 3D DataArray
-    is_ibound: bool, optional
-        If True, ``da`` is interpreted as ibound, with values 0: inactive, 1: active, -1 constant head.
-        If False, ``upper_active_layer`` is interpreted as first layer that has data.
-        Default is True.
-    include_constant_head : bool, optional
-        If True and ``is_ibound``, also include constant head cells.
-        Default is False.
-
-    Returns
-    -------
-    2d xr.DataArray of layernumber of upper active model layer
+    active: {DataArray, UgridDataArray}
+        Grid of booleans designating active cell.
     """
-    da = da.load()
-    if is_ibound:
-        # check if indeed ibound: convertible to int
-        if not da.astype(int).equals(da):
-            raise ValueError(
-                "Passed DataArray is no ibound, while is_bound was set to True"
-            )
-        # include constant head cells (?)
-        if include_constant_head:
-            is_active = da.fillna(0) != 0  # must be filled for argmax
-        else:
-            is_active = da.fillna(0) > 0
-    else:
-        is_active = ~da.isnull()
+    layer = active.coords["layer"]
+    # Set nodata to sentinel value to prevent a dtype shift from integer to
+    # float as np.nan forces float.
+    nodata = layer.max() + 1
+    return layer.where(active, nodata).min("layer")
 
-    # get layer of upper active cell
-    da = is_active.layer.isel(layer=is_active.argmax(dim="layer"))
-    da = da.drop_vars("layer")
 
-    # skip where no active cells
-    return da.where(is_active.sum(dim="layer") > 0)
+@preserve_gridtype
+def get_upper_active_grid_cells(active: GridDataArray) -> GridDataArray:
+    """
+    Returns grid of booleans designating location of the uppermost active grid
+    cell.
+
+    Parameters
+    ----------
+    active: {DataArray, UgridDataArray}
+        Grid of booleans designating active cell.
+    """
+    layer = active.coords["layer"]
+    upper_active_layer = get_upper_active_layer_number(active)
+    return layer == upper_active_layer
+
+
+@preserve_gridtype
+def get_lower_active_layer_number(active: GridDataArray) -> GridDataArray:
+    """
+    Returns two-dimensional grid of integers with the layer number of the lower
+    most active cell.
+
+    Parameters
+    ----------
+    active: {DataArray, UgridDataArray}
+        Grid of booleans designating active cell.
+    """
+    layer = active.coords["layer"]
+    # Set nodata to sentinel value to prevent a dtype shift from integer to
+    # float as np.nan forces float.
+    nodata = layer.min() - 1
+    return layer.where(active, nodata).max("layer")
+
+
+@preserve_gridtype
+def get_lower_active_grid_cells(active: GridDataArray) -> GridDataArray:
+    """
+    Returns grid of booleans designating location of the lowermost active grid
+    cell.
+
+    Parameters
+    ----------
+    active: {DataArray, UgridDataArray}
+        Grid of booleans designating active cell.
+    """
+    layer = active.coords["layer"]
+    lower_active_layer = get_lower_active_layer_number(active)
+    return layer == lower_active_layer
